refactor: Move OverviewCard to separate file
The OverviewCard composable has been moved from `HomeScreen.kt` to its own file `composable/OverviewCard.kt` to improve code organization.

Additionally, a size modifier has been added to the FloatingActionButton in the `TudeeButton` preview.
